# project/docs/swagger.yaml
basePath: /
definitions:
  model.Collection:
    properties:
      config:
        allOf:
        - $ref: '#/definitions/model.CollectionConfig'
        description: Настройки генерации
      created_at:
        type: string
      description:
        example: User data collection
        type: string
      fields:
        description: Шаблоны полей для генерации
        items:
          $ref: '#/definitions/model.FieldTemplate'
        type: array
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        example: users
        type: string
      project_id:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
  model.CollectionConfig:
    properties:
      count:
        description: Количество генерируемых записей
        example: 10
        type: integer
      seed:
        description: Seed для воспроизводимости
        type: integer
    type: object
  model.CreateCollectionRequest:
    properties:
      config:
        allOf:
        - $ref: '#/definitions/model.CollectionConfig'
        description: Опционально при создании
      description:
        example: User data
        maxLength: 500
        type: string
      fields:
        description: Опционально при создании
        items:
          $ref: '#/definitions/model.FieldTemplate'
        type: array
      name:
        example: users
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    type: object
  model.CreateProjectRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
  model.FieldTemplate:
    properties:
      format:
        example: email
        type: string
      max:
        description: для numbers
        example: 100
        type: number
      min:
        description: для numbers
        example: 0
        type: number
      name:
        example: email
        type: string
      options:
        description: для enum
        example:
        - active
        - inactive
        items:
          type: string
        type: array
      required:
        example: true
        type: boolean
      type:
        enum:
        - string
        - number
        - boolean
        - date
        example: string
        type: string
      unique:
        example: false
        type: boolean
    type: object
  model.Project:
    properties:
      api_key:
        description: Уникальный ключ для доступа к API проекта
        type: string
      base_url:
        description: https://<api_key>.mockingcode.org
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.UpdateCollectionRequest:
    properties:
      config:
        allOf:
        - $ref: '#/definitions/model.CollectionConfig'
        description: Можно обновить конфиг
      description:
        example: Updated user data
        maxLength: 500
        type: string
      fields:
        description: Можно обновить поля
        items:
          $ref: '#/definitions/model.FieldTemplate'
        type: array
      is_active:
        example: true
        type: boolean
      name:
        example: updated-users
        maxLength: 50
        type: string
    type: object
  model.UpdateProjectRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
host: localhost:8082
info:
  contact:
    email: support@mockingcode.dev
    name: API Support
    url: http://mockingcode.dev/support
  description: Project management service for MockingCode platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://mockingcode.dev/terms/
  title: MockingCode Project Service API
  version: "1.0"
paths:
  /api-keys/{apiKey}/validate:
    get:
      description: Validate API Key and return project information
      parameters:
      - description: API Key
        in: path
        name: apiKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Validate API Key
      tags:
      - api-keys
  /projects:
    get:
      description: Get list of all projects for authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project for authenticated user
      parameters:
      - description: Project data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Delete project and all its collections
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Get specific project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update project name and description
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
  /projects/{id}/collections:
    get:
      consumes:
      - application/json
      description: Get list of collections for a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Collection'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get collections for a project
      tags:
      - collections
    post:
      consumes:
      - application/json
      description: Create a new collection
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Collection data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateCollectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Collection'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create collections for a project
      tags:
      - collections
  /projects/{projectId}/collections/{collectionId}:
    delete:
      description: Delete collection
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete collection
      tags:
      - collections
    get:
      description: Get specific collection by ID
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get collection by ID
      tags:
      - collections
    put:
      consumes:
      - application/json
      description: Update collection name, description, fields or config
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        type: integer
      - description: Collection data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateCollectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update collection
      tags:
      - collections
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"